Change parameter to constant in PSROIPooling

+++ inference-engine/tests/functional/shared_test_classes/include/shared_test_classes/single_layer/psroi_pooling.hpp
@@ -18,7 +18,7 @@
 namespace LayerTestsDefinitions {
 
 using psroiParams = std::tuple<std::vector<size_t>,            // input shape
-                               std::vector<size_t>,            // coords shape
+                               std::vector<float>,             // coords value
                                size_t,                         // output_dim
                                size_t,                         // group_size
                                float,                          // Spatial scale
